pipeline_JL/AI_Data_Explorer/requirements.txt
openai
python-dotenv
flask
flask-cors
pandas
